Update TabletFile to use getters for different situations
* Created methods for read/insert/updateDel and have them called
throughout the code for the appropriate situation
* Throw exception if we try to get metadata for update or delete and one
does not exist
* Call inserted() method on TabletFile to update metadata entry when a
file is new and inserted
* Drop FileUtil.toPathStrings() and use object collections instead
* Replace String path with TabletFile in FileUtil

+++ server/base/src/main/java/org/apache/accumulo/server/util/MasterMetadataUtil.java
@@ -88,7 +88,7 @@ public class MasterMetadataUtil {
 
     for (Entry<Long,? extends Collection<TabletFile>> entry : bulkLoadedFiles.entrySet()) {
       for (TabletFile ref : entry.getValue()) {
-        tablet.putBulkFile(ref, entry.getKey().longValue());
+        tablet.putBulkFile(ref, entry.getKey());
       }
     }
